# openapiv2/laptop_service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "laptop_service.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "LaptopService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/laptop/create": {
      "post": {
        "operationId": "LaptopService_CreateLaptop",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pcbookCreateLaptopResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pcbookCreateLaptopRequest"
            }
          }
        ],
        "tags": [
          "LaptopService"
        ]
      }
    },
    "/v1/laptop/rate": {
      "post": {
        "operationId": "LaptopService_RateLaptop",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/pcbookRateLaptopResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of pcbookRateLaptopResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": " (streaming inputs)",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pcbookRateLaptopRequest"
            }
          }
        ],
        "tags": [
          "LaptopService"
        ]
      }
    },
    "/v1/laptop/search": {
      "get": {
        "operationId": "LaptopService_SearchLaptop",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/pcbookSearchLaptopResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of pcbookSearchLaptopResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "filter.maxPriceUsd",
            "in": "query",
            "required": false,
            "type": "number",
            "format": "double"
          },
          {
            "name": "filter.minCpuCores",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "filter.minCpuGhz",
            "in": "query",
            "required": false,
            "type": "number",
            "format": "double"
          },
          {
            "name": "filter.minRam.value",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "filter.minRam.unit",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "UNSPECIFIED",
              "BIT",
              "BYTE",
              "KILLOBYTE",
              "MEGABYTE",
              "GIGABYTE",
              "TERABYTE"
            ],
            "default": "UNSPECIFIED"
          }
        ],
        "tags": [
          "LaptopService"
        ]
      }
    },
    "/v1/laptop/upload_image": {
      "post": {
        "operationId": "LaptopService_UploadImage",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pcbookUploadImageResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": " (streaming inputs)",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pcbookUploadImageRequest"
            }
          }
        ],
        "tags": [
          "LaptopService"
        ]
      }
    }
  },
  "definitions": {
    "KeyboardLayout": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "QWERTY",
        "QWERTZ",
        "AZERTY"
      ],
      "default": "UNKNOWN"
    },
    "MemoryUnit": {
      "type": "string",
      "enum": [
        "UNSPECIFIED",
        "BIT",
        "BYTE",
        "KILLOBYTE",
        "MEGABYTE",
        "GIGABYTE",
        "TERABYTE"
      ],
      "default": "UNSPECIFIED"
    },
    "ScreenPanel": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "IPS",
        "OLED"
      ],
      "default": "UNKNOWN"
    },
    "ScreenResolution": {
      "type": "object",
      "properties": {
        "width": {
          "type": "integer",
          "format": "int64"
        },
        "height": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "StorageDriver": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "HDD",
        "SSD"
      ],
      "default": "UNKNOWN"
    },
    "pcbookCPU": {
      "type": "object",
      "properties": {
        "brand": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "numCores": {
          "type": "integer",
          "format": "int64"
        },
        "numThreads": {
          "type": "integer",
          "format": "int64"
        },
        "minGhz": {
          "type": "number",
          "format": "double"
        },
        "maxGhz": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "pcbookCreateLaptopRequest": {
      "type": "object",
      "properties": {
        "laptop": {
          "$ref": "#/definitions/pcbookLaptop"
        }
      }
    },
    "pcbookCreateLaptopResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "pcbookGPU": {
      "type": "object",
      "properties": {
        "brand": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "numCores": {
          "type": "integer",
          "format": "int64"
        },
        "numThreads": {
          "type": "integer",
          "format": "int64"
        },
        "minGhz": {
          "type": "number",
          "format": "double"
        },
        "maxGhz": {
          "type": "number",
          "format": "double"
        },
        "memory": {
          "$ref": "#/definitions/pcbookMemory",
          "title": "memory in GB"
        }
      }
    },
    "pcbookImageInfo": {
      "type": "object",
      "properties": {
        "laptopId": {
          "type": "string"
        },
        "imageType": {
          "type": "string"
        }
      }
    },
    "pcbookKeyboard": {
      "type": "object",
      "properties": {
        "layout": {
          "$ref": "#/definitions/KeyboardLayout"
        },
        "backlit": {
          "type": "boolean"
        }
      }
    },
    "pcbookLaptop": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "brand": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "cpu": {
          "$ref": "#/definitions/pcbookCPU"
        },
        "ram": {
          "$ref": "#/definitions/pcbookMemory"
        },
        "gpu": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/pcbookGPU"
          }
        },
        "storage": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/pcbookStorage"
          }
        },
        "screen": {
          "$ref": "#/definitions/pcbookScreen"
        },
        "keyboard": {
          "$ref": "#/definitions/pcbookKeyboard"
        },
        "weightKg": {
          "type": "number",
          "format": "float"
        },
        "weightLbs": {
          "type": "number",
          "format": "float"
        },
        "priceUsd": {
          "type": "number",
          "format": "double"
        },
        "releaseYear": {
          "type": "integer",
          "format": "int64"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "pcbookLaptopFilter": {
      "type": "object",
      "properties": {
        "maxPriceUsd": {
          "type": "number",
          "format": "double"
        },
        "minCpuCores": {
          "type": "integer",
          "format": "int64"
        },
        "minCpuGhz": {
          "type": "number",
          "format": "double"
        },
        "minRam": {
          "$ref": "#/definitions/pcbookMemory"
        }
      }
    },
    "pcbookMemory": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string",
          "format": "uint64"
        },
        "unit": {
          "$ref": "#/definitions/MemoryUnit"
        }
      }
    },
    "pcbookRateLaptopRequest": {
      "type": "object",
      "properties": {
        "laptopId": {
          "type": "string"
        },
        "score": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "pcbookRateLaptopResponse": {
      "type": "object",
      "properties": {
        "laptopId": {
          "type": "string"
        },
        "ratedCount": {
          "type": "integer",
          "format": "int64"
        },
        "averageScore": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "pcbookScreen": {
      "type": "object",
      "properties": {
        "sizeInch": {
          "type": "number",
          "format": "float"
        },
        "resolution": {
          "$ref": "#/definitions/ScreenResolution"
        },
        "panel": {
          "$ref": "#/definitions/ScreenPanel"
        },
        "multiTouch": {
          "type": "boolean"
        }
      }
    },
    "pcbookSearchLaptopResponse": {
      "type": "object",
      "properties": {
        "laptop": {
          "$ref": "#/definitions/pcbookLaptop"
        }
      }
    },
    "pcbookStorage": {
      "type": "object",
      "properties": {
        "driver": {
          "$ref": "#/definitions/StorageDriver"
        },
        "memory": {
          "$ref": "#/definitions/pcbookMemory"
        }
      }
    },
    "pcbookUploadImageRequest": {
      "type": "object",
      "properties": {
        "info": {
          "$ref": "#/definitions/pcbookImageInfo"
        },
        "chunkData": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "pcbookUploadImageResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "size": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}